Disabled buttons that lead to blank screens
Some buttons in both MainActivityFragment.java and HomeScreen.java have
been disabled because they lead to a blank screen. A Toast message pops
up instead.

The following buttons are disabled :
getHelpButton, reportingConfidentiality, reportingMoreResources, and
safetyResourceButton

Comrade numbers are no longer hard-coded

Comrade numbers are now stored
in a SharedPreference as opposed to being
hard-coded in an array.

Also,
the Toast message that is sent after submission is no longer
hard-coded.
It now displays the correct number of comrades that your
text message
was sent to.

Comrade numbers are now editable after submission

The comrade numbers
do not permanently disappear from the EditTexts
immediately after
exiting the Activity, so they are easily editable
after they are
submitted.